Created standard rule set for Game of Life. Pair programmed by Michael and Geoff.
We have created a standard rule set for the Game of Life and tested it.
The standard rule set means that a cell is born when it has exactly
three living neighbours and that a cell stays alive when it has two or
three living neighbours. Feature #11 can now be closed.