GameOfLife class began. Pair programmed by Michael and Geoff.
Related Issues: Feature #13
				Support #15
				
We began work on the GameOfLife class that will represent the game
engine as a whole, but ran into issues with how the correct RuleSet
object will be assigned to it. Details can be found on redmine, under
Issue #15.
We began writing the tests for the class, but only some of them are
completed since we ran out of time.

This should not have been pushed, but everyone must have the code
available to properly review it.