Wrote the GameOfLife class. Paired by Michael and Geoff.
Completed the GameOfLife class. There is one failing test
(test_next_turn). It fails because GameOfLife._calculate_next_generation
currently returns a collection of cells as opposed to a grid. It cannot
currently create a grid because a grid requires a size and the
GameOfLife class has no idea of this. Discuss.
A further issue remains of whether a grid object can be passed into the
GameOfLife ctor - this means that other classes have access to Grid
objects. Not sure whether that is a good idea. (Feature #13)