More work done on the GameOfLife class. Paired by Michael and Geoff.
In trying to fix the failed test in the framework for GameOfLife, we
have refactored Grid to be able to take two optional parameters. One is
the size of the grid, as a tuple, and one is a pattern of cells. The
current issue is that at the moment, when a grid object is initialised,
it will always be a 10x10 grid of dead cells. (Feature #13).