Automated test for HTTP access to the Pattern Input JSON Receiver view. Paired by Richard and Michael.
Testing of the view over HTTP has been automated, with a framework for further HTTP testing
later on in the project also added. Additionally, the MAKEFILE was edited so that the command
make run now works - this sets up the server in development mode and reload.

All tests now pass (Support #106).