Template for the web application created. Paired by Michael and Richard.
A template for the web application has been created, providing us with parts of the website
that are unchanged on each page. So far this includes a ProjectConway logo and a navbar.
The navbar has been programmed so that on 'extra-small' devices, it becomes a dropdown
menu.

All tests succeed (Feature #87)