Made progress on the pattern input grid
Changed from using canvas calls to using the jCanvas library.
Fixed some of the maths.
Worked out how to draw single pixel lines in canvas (why on earth is this so differecult? Why doesn't a line width of 1 mean 1 pixel?)

Next the grid needs to be update to make it responsive.