Implemented dialogs for the pattern creation page. Paired by Michael and Richard.
The dialogs for the pattern creation page were implemented using modals from BootstrapJS. The first was a loading dialog, used while the game engine calculates the length of a pattern; the second was a success dialog that tells the user both how many generations a pattern will last and how much time (in seconds) it will last for.

Additionally, an error dialog appears if, for any reason, the pattern AJAX fails.

All testing has been completed (Feature #102).