Fixed the  invisable alert bug.
An invisable alert was covering up a portion of the grid, so that it couldn't be clicked. This was especially impacting mobiles.
I have updated it so that the alert starts of the screen and is brought onto the screen when the issue occurs. Also had to update the closing button.