Datepicker created. Paired by Michael and Richard.
A Datepicker has been put on the scheduling page so that people can choose the date on which they will view their pattern. This has also been made to point to the current date should no date already be saved to the session. The size of the datepicker was also changed so that it fits the style of the page. (Feature #90)