Tested the scheduler's ability to communicate with the database. Paired by Richard and Michael.
We have finished writing tests for the scheduler, and currently these tests fail because the tested
methods lack implementation. Among these tests however, is a test of whether the scheduler is capable
of properly connecting with the database, and we have discovered it can both push and pull information
with ease. (Feature #90)