Implemented the view without error checking. Paired by Richard and Michael.
The functionality of the view has been implemented but there is not yet any code in place for
checking whether the chosen date is too far in the past or too far in the future. We have switched
to postgreSQL, but it is currently hardcoded completely to Richards machine and that needs to be
fixed. (Feature #90)