More filler added. Programmed by Michael.
I added some more filler to the home page so that it gave a better impression
of what it would look like. Some of this text will be replaced by pictures,
but if we can keep it on one page, without scrolling, I think we are heading
in the right direction. (Feature #87)