Better screensavers in the screensavers.txt file. Paired by Michael and Richard.
We have added some better screensavers to the internal database so screensaver mode has more variety from
which to choose. Some of these patterns end in still-lives, or don't last the whole five minutes, but we think this
is more a product of the 10x10 than of the patterns we have chosen.  (Support #197)