Rewrote some of the javascript for the scheduler page. Programmed by Michael and Richard.
We began rewriting the javascript for the scheduling page to use the new functionality of the server-side database. Instead of doing a request each time the user selects a new date and hour, it has been limited to one request per date selection. This
pulls every available time slot for the day and stores them to be repainted into the page when the user changes the hour. (Feature #140)