Fixed some Javascript and some views. Paired by Michael and Richard.
The scheduling page is now in a state where selecting a date from the datepicker will pull the available time slots for a day and store it in a js var, to be used in updating the viewing hour and viewing slot. A main issue with this was the asynchronous nature
of the js .ajax call, but now it works. Not all unit tests pass, but it has been decided to fully implement the changes before fixing this issue, so we can work out what is a broken test and what is our broken code. A view has also been fixed to impliment
 the earlier changes made to the server-side database. (Feature #140)