The bug in PySerial has been fixed. Paired by Michael and Richard.
The bug in PySerial has now been permanently fixed by us hosting it in the DisplayAdapter package. This means the change we had been making multiple times is now permanently written into our code and must not be done time and time again.
We have tested this on the prototype display and found it to work. (Bug #207)